\name{bitsgHs}
\alias{bitsgHs}

\title{Internal Function}

\description{It provides the quantites needed to calculate the log-likelihood, gradient and Hessian matrix for 
penalized or unpenalized maximum likelihood optimization, for a number of copula models.}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}
