\name{copulaBitsD}
\alias{copulaBitsD}

\title{Internal Function}

\description{It provides first and second derivatives for copulas with respect to their margins and the association parameter \eqn{\theta}{\theta}.

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}










