\name{fit.SemiParSampleSel}
\alias{fit.SemiParSampleSel}

\title{Internal Function}

\description{Performs the optimization (using Newton-Raphson's method) and automatic smoothing parameter selection.
}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}








