\name{fit.SemiParSampleSel1}
\alias{fit.SemiParSampleSel1}

\title{Internal Function}

\description{It performs the optimization using a trust region algorithm as well as automatic smoothing parameter estimation. This
is more efficient and reliable.}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}








