\name{ghss}
\alias{ghss}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and hessian matrix for penalized or unpenalized maximum likelihood optimization.}



\author{

Maintainer: Giampiero Marra \email{giampiero@stats.ucl.ac.uk}

}



\references{

Marra G. and Radice R. (submitted), Semiparametric Continuous Response Models for Controlling Sample Selection Bias.

}









