\name{ghss2}
\alias{ghss2}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and Hessian matrix for unpenalized maximum likelihood optimization for 
the standard deviation and association parameters only, for several bivariate copula distributions.}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}






