\name{ghssAMH}
\alias{ghssAMH}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and Hessian matrix for penalized or unpenalized maximum 
likelihood optimization when using the Ali-Mikhail-Haq copula with normal margins.}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}








