\name{ghssAMH1}
\alias{ghssAMH1}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and Hessian matrix for penalized maximum likelihood optimization 
of all parameters but the standard deviation and association coefficient, when using the Ali-Mikhail-Haq copula with normal margins.}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}








