\name{ghssAMH1g}
\alias{ghssAMH1g}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and Hessian matrix for penalized maximum likelihood optimization 
of all parameters but the dispersion and association coefficients, when using the Ali-Mikhail-Haq copula with normal and gamma margins.}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}








