\name{ghssF}
\alias{ghssF}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and Hessian matrix for penalized or unpenalized maximum 
likelihood optimization when using the Frank copula with normal margins.}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}







