\name{ghssF1}
\alias{ghssF1}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and Hessian matrix for penalized maximum likelihood optimization 
of all parameters but the standard deviation and association coefficient, when using the Frank copula with normal margins.}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}










