\name{ghssF1g}
\alias{ghssF1g}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and Hessian matrix for penalized maximum likelihood optimization 
of all parameters but the dispersion and association coefficients, when using the Frank copula with normal and gamma marginals.}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}










