\name{ghssF2}
\alias{ghssF2}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and Hessian matrix for unpenalized maximum likelihood optimization for 
the standard deviation and association parameters only, when using the Frank copula with normal margins.}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}










