\name{ghssF2g}
\alias{ghssF2g}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and Hessian matrix for unpenalized maximum likelihood optimization for 
the dispersion and association parameters only, when using the Frank copula with normal and gamma marginals.}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}










