\name{ghssG2g}
\alias{ghssG2g}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and Hessian matrix for unpenalized maximum likelihood optimization for 
the dispersion and association parameters only, when using the Gumbel copula with normal and gamma margins.}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}






