\name{ghssGg}
\alias{ghssGg}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and Hessian matrix for penalized or unpenalized maximum 
likelihood optimization when using the Gumbel copula with normal and gamma margins.}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}








