\name{ghssJ1}
\alias{ghssJ1}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and Hessian matrix for penalized maximum likelihood optimization 
of all parameters but the standard deviation and association coefficient, when using the Joe copula with normal margins.}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}




