\name{ghssJg}
\alias{ghssJg}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and Hessian matrix for 
penalized or unpenalized maximum likelihood optimization when using the Joe copula 
with normal and gamma marginals.}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}



