\name{ghssN2}
\alias{ghssN2}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and Hessian matrix for unpenalized maximum likelihood optimization for 
the standard deviation and correlation parameters only,
when assuming bivariate normal errors.}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}



\references{

Marra G. and Radice R. (in press), Estimation of a Regression Spline Sample Selection Model. \emph{Computational Statistics and Data Analysis}.

}









