\name{print.aver}
\alias{print.aver}

\title{Print an aver object}

\description{The print method for a \code{aver} object.}

\usage{


\method{print}{aver}(x, ...)


}


\arguments{ 


\item{x}{A \code{aver} object produced by \code{aver()}.}

\item{...}{Other arguments.}

}


\details{ 

\code{print.aver} prints the lower confidence interval limit, estimated average and upper confidence interval limit. 

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}




\seealso{
\code{\link{aver}}
}

\keyword{semiparametric sample selection modelling} 









