\name{print.est.aver}
\alias{print.est.aver}

\title{Print an est.aver object}

\description{The print method for a \code{est.aver} object.}

\usage{


\method{print}{est.aver}(x, ...)


}


\arguments{ 


\item{x}{A \code{est.aver} object produced by \code{est.aver()}.}

\item{...}{Other arguments.}

}


\details{ 

\code{print.est.aver} prints the lower confidence interval limit, estimated average and upper confidence interval limit. 

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}




\seealso{
\code{\link{est.aver}}
}

\keyword{semiparametric sample selection modelling} 









