\name{print.SemiParSampleSel}
\alias{print.SemiParSampleSel}

\title{Print a SemiParSampleSel object}

\description{The print method for a \code{SemiParSampleSel} object.}

\usage{


\method{print}{SemiParSampleSel}(x, ...)


}


\arguments{ 


\item{x}{A \code{SemiParSampleSel} object produced by \code{SemiParSampleSel()}.}

\item{...}{Other arguments.}

}


\details{ 

\code{print.SemiParSampleSel} prints out the family, model equations, total number of observations, estimated association, spread 
and shape (if present) coefficients, and total effective degrees of freedom for the penalized or unpenalized model. 

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}




\seealso{
\code{\link{SemiParSampleSel}}
}


\keyword{semiparametric sample selection modelling} 








