\name{print.summary.SemiParSampleSel}
\alias{print.summary.SemiParSampleSel}

\title{Print a summary.SemiParSampleSel object}

\description{The print method for a \code{summary.SemiParSampleSel} object.}

\usage{


\method{print}{summary.SemiParSampleSel}(x, digits = max(3, getOption("digits") - 3), 
           signif.stars = getOption("show.signif.stars"), ...)


}


\arguments{ 


\item{x}{A \code{summary.SemiParSampleSel} object produced by \code{summary.SemiParSampleSel()}.}

\item{digits}{Number of digits printed in output.} 

\item{signif.stars}{By default significance stars are printed alongside output.} 

\item{...}{Other arguments.}

}


\details{ 

\code{print.summary.SemiParSampleSel} prints model term summaries. 

}




\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}



\seealso{
\code{\link{summary.SemiParSampleSel}}
}


\keyword{semiparametric sample selection modelling} 








