\name{SOEP}
\alias{SOEP}

\title{SOEP - German labour mobility data}

\description{The data set orginates from the German Socio-Economic Panel survey of 1984. It contains characteristics of 2651 individuals such as employment, marital status, education and political preferences. Note that the data set is significantly smaller than the original data and any source of identification (e.g., ID number) has been removed. }


\usage{

data(SOEP)

}



\references{


Wyszynski K. and Marra G. (submitted), Sample selection model for count data: a tutorial in R using the package SemiParSampleSel.

}






