\name{ss.checks}
\alias{ss.checks}

\title{Some convergence diagnostics}

\description{It takes a fitted \code{SemiParSampleSel} object produced by \code{SemiParSampleSel()} and produces some diagnostic information about the fitting procedure.

}

\usage{

ss.checks(x)


}


\arguments{ 


\item{x}{A \code{SemiParSampleSel} object produced by \code{SemiParSampleSel()}.}


}




\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}



\seealso{
\code{\link{SemiParSampleSel}}
}


\keyword{semiparametric sample selection modelling} 








