\name{st.theta.star}
\alias{st.theta.star}

\title{Internal Function}

\description{It computes a starting value for dependence parameter \eqn{\theta}{\theta} and transforms it depending on the copula used.}


\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}








