\name{theta.tau}
\alias{theta.tau}

\title{Internal Function}

\description{ Given an estimated value of theta.star this function computes theta using a proper transformation, and Kendall's tau for a given copula. }



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}








