\name{fit.SemiParSampleSel}
\alias{fit.SemiParSampleSel}

\title{Internal Function}

\description{It performs the optimization using a trust region algorithm as well as automatic smoothing parameter estimation.
}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}








