\name{ghss}
\alias{ghss}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and Hessian matrix for penalized or unpenalized maximum 
likelihood optimization for several bivariate copula distributions.

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}










