\name{ghss1}
\alias{ghss1}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and hessian matrix for penalized maximum likelihood optimization 
of all parameters except for the standard deviation and correlation.}



\author{

Maintainer: Giampiero Marra \email{giampiero@stats.ucl.ac.uk}

}



\references{

Marra G. and Radice R. (submitted), Semiparametric Continuous Response Models for Controlling Sample Selection Bias.

}









