\name{ghss2}
\alias{ghss2}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and hessian matrix for unpenalized maximum likelihood optimization for 
the standard deviation and correlation parameters only.}



\author{

Maintainer: Giampiero Marra \email{giampiero@stats.ucl.ac.uk}

}



\references{

Marra G. and Radice R. (submitted), Semiparametric Continuous Response Models for Controlling Sample Selection Bias.

}









