\name{theta.tau}
\alias{theta.tau}

\title{Internal Function}

\description{Given an estimated value of \eqn{\theta^{*}}{\theta^{*}}, this function computes \eqn{\theta}{\theta} using a proper 
transformation, and Kendall's \eqn{\tau}{\tau} for a given copula. }



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}








