\name{working.comp}
\alias{working.comp}

\title{Internal Function}

\description{It efficiently calculates the working model quantities needed to implement the automatic multiple smoothing parameter 
procedure by exploiting the band structure of the weight matrix.}




\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}



\references{

Marra G. and Radice R. (2013), Estimation of a Regression Spline Sample Selection Model. \emph{Computational Statistics and Data Analysis}, 61, 158-173.

Wojtys M., Marra G. and Radice R. (submitted), Copula Regression Spline Sample Selection Models: The R Package SemiParSampleSel.

}






