\name{working.comp}
\alias{working.comp}

\title{Internal Function}

\description{It efficiently calculates the working model quantities needed to implement the automatic multiple smoothing parameter estimation 
procedure by exploiting a result which leads to very fast and stable calculations.}




\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}



\references{

Wojtys M., Marra G. and Radice R. (in press), Copula Regression Spline Sample Selection Models: The R Package SemiParSampleSel. \emph{Journal of Statistical Software}.

}






